feature: 1. 新增LogCommand和CheckCommand

+++ common/src/main/java/me/youfang/common/utils/IOExt.kt
@@ -1,9 +1,14 @@
 package me.youfang.common.utils
 
+import org.apache.commons.io.FileExistsException
 import org.apache.commons.io.FileUtils
+import org.apache.commons.io.IOUtils
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
+import java.io.IOException
+import java.io.InputStream
+import java.io.OutputStream
 import java.nio.file.CopyOption
 import java.nio.file.StandardCopyOption
 import java.util.UUID
@@ -18,6 +23,16 @@ fun String.writeToFileWithTempFile(file: File) {
     writeToFile(tempFile)
     if (file.exists()) {
         file.delete()
+    }
+    FileUtils.moveFile(tempFile, file)
+}
+
+fun InputStream.writeToFileWithTempFile(file: File) {
+    if (file.exists()) throw FileExistsException("文件已存在: $file")
+    val tempFile = File(file.parentFile, UUID.randomUUID().toString())
+    FileUtils.forceMkdirParent(tempFile)
+    FileOutputStream(tempFile).use {
+        copyTo(it)
     }
     FileUtils.moveFile(tempFile, file)
 }
